+++ IO.Executable/Program.cs
@@ -1,0 +1,7 @@
+﻿using static Messerli.IO.ConsoleUtility;
+
+if (OperatingSystem.IsWindows() && IsLastProcessAttachedToConsole())
+{
+    Console.WriteLine("Press Enter to continue...");
+    Console.ReadLine();
+}