+++ changelog.md
@@ -1,4 +1,8 @@
 # Changelog
+## 0.2.1
+* Added `ConsoleUtility.IsLastProcessAttachedToConsole` for determining
+  if the current process is the last one attached to the console.
+
 ## 0.2.0
 * Added `Retrier.RetryWhileFileIsInUseAsync`.
 * Removed `StreamExtension`. The source code of the 0.1.0 version has been moved to [io-old](https://github.com/messerli-informatik-ag/io-old).